Avoid the use of mmap in cmime_message.c when building on windows.
The existing code mmaps a file, then callocs some space and
copies the file contents into it.

Why not just calloc the space, and fread from the file?

Windows can't mmap. If it was me, I'd avoid mmap entirely as
it doesn't seem to be gaining anything here, but that's not
my decision!